Server Microservice_with3 models working/requirements.txt
azure-cognitiveservices-speech==1.36.0
librosa==0.10.1
Flask==3.0.2
h5py==3.10.0
huggingface-hub==0.21.4
joblib==1.3.2
keras==3.0.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
matplotlib-inline==0.1.6
noisereduce==3.0.2
numba==0.59.0
numpy==1.26.4
pandas==2.2.1
pydub==0.25.1
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
tensorflow==2.16.1
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.38.2